requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.1.3
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycparser==2.21
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
Werkzeug==2.1.2
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
zipp==3.8.0
